+++ Modules/Intent.Modules.Integration.HttpClients/release-notes.md
@@ -1,7 +1,7 @@
 ﻿### Version 5.1.3
 
 - Improvement: Added support for gathering Service Proxy information from the Services designer too.
-- Fixed: Added setting to either make use of Pass-through Authorization Headers or Token Management Servers (both cannot be used at the same time per http client).
+- Fixed and Improved: Added setting to either make use of Transmittable Access Token or Client Access Token Management (both cannot be used at the same time per http client).
 
 ### Version 5.1.2